requirements/prod.txt
## The following requirements were added by pip freeze:
boto3==1.18.16
botocore==1.21.16
certifi==2021.5.30
charset-normalizer==2.0.4
falcon==3.0.0
falcon-auth==1.1.0
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
passlib==1.7.4
protobuf==3.17.3
PyJWT==2.1.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
